Refactor tests before writing tests for ignored cases
