+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/dependencyInExportDefaultSubgraph/sourceCodes/entryPoint.ts
