+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/dependencyInExportDefaultSubgraph/sourceCodes/frameworkGraph.ts
