+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/dependencyInSubgraph/index.ts
@@ -1,4 +1,4 @@
-import { TestCase } from "../..";
+import { TestCase } from "../../..";
 import * as path from 'path';
 
 export default {
