+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/dependencyInSubgraph/sourceCodes/frameworkGraph.ts
