+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedClass/index.ts
@@ -1,22 +1,23 @@
-import { TestCase } from "../..";
+import { TestCase } from "../../..";
 import * as path from 'path';
 
+
 export default {
-  name: 'injected hook dependenciesOf type alias',
+  name: 'injected class',
   entryPoint: path.resolve(__dirname, './sourceCodes/entryPoint.ts'),
   position: {
-    line: 9,
-    character: 29
+    line: 10,
+    character: 33
   },
   result: {
-    uri: path.resolve(__dirname, './sourceCodes/entryViewModel.ts'),
+    uri: path.resolve(__dirname, './sourceCodes/gameModel.ts'),
     range: {
       start: {
-        line: 3,
-        character: 13
+        line: 2,
+        character: 0
       },
       end: {
-        line: 7,
+        line: 4,
         character: 1
       }
     }
