+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedClass/sourceCodes/entryPoint.ts
