+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedClassDependenciesOfTypeAlias/sourceCodes/gameGraph.ts
