+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedExportDefaultClass/sourceCodes/entryPoint.ts
