+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedExportDefaultClass/sourceCodes/gameGraph.ts
