+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedExportDefaultGraph/sourceCodes/entryGraph.ts
