+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedHook/sourceCodes/component.ts
