+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedHookDependenciesOfTypeAlias/sourceCodes/entryViewModel.ts
