+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedHookDifferentInjectedTypeName/sourceCodes/component.ts
