+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedHookDifferentInjectedTypeName/sourceCodes/entryViewModel.ts
