+++ packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/handled/injectedHookTypedProvider/sourceCodes/entryViewModel.ts
