--- packages/vscode-language-server-obsidian/server/tests/integration/goToDefinition/injectedClass/index.ts
@@ -1,25 +1,0 @@
-import { TestCase } from "../..";
-import * as path from 'path';
-
-
-export default {
-  name: 'injected class',
-  entryPoint: path.resolve(__dirname, './sourceCodes/entryPoint.ts'),
-  position: {
-    line: 10,
-    character: 33
-  },
-  result: {
-    uri: path.resolve(__dirname, './sourceCodes/gameModel.ts'),
-    range: {
-      start: {
-        line: 2,
-        character: 0
-      },
-      end: {
-        line: 4,
-        character: 1
-      }
-    }
-  }
-} satisfies TestCase;
